Ask OS to kill us on pipe error
Rastair should be polite and accept its output going to stdout which could be
head or less that the user quits before reading the full stream. Before this
commit this causes errors like

    Error:
       0: Failed to convert VCF to BED
       1: Failed to write record
       2: Failed to write record for chr12:55009199
       3: Broken pipe (os error 32)

Of course, I was ready to add

    .map_err(|err| {
        if let Some(io_err) = err.root_cause().downcast_ref::<std::io::Error>()
            && io_err.kind() == ErrorKind::BrokenPipe
        {
            // Exit gracefully if the output pipe is broken
            // (e.g., when piping to `head` or `less`).
            // 128 + 13 = 141 is the usual exit code for SIGPIPE
            std::process::exit(141);
        } else {
            err
        }
    })

in the right places but then I saw that I won’t get any useful errors from
htslib – it has one generic Error::WriteRecord only.

Stealing the idea from rastair1, I’ll go with resetting the signal handler and
letting the kernel kill the program instead. This is not optimal as we might
also write a BED file at the same time, which might need to be flushed, but I’ll
take that chance. Most people who want complete files will not pipe it to head.

Fixes #17