Vcf->Bed: Redo some data fetching to fix bufferse
If I understand the [docs] correctly, the lifetimes of the buffer I get
from `record.info().float()` pretty are pretty much a lie. I need to
keep that buffer in scope to access the data and nothing in the types
will make sure I do that.

But okay, using if let chains works in keeping it in scope.

[docs]: https://docs.rs/rust-htslib/0.49.0/rust_htslib/bcf/record/struct.Format.html#method.float