Refactor fetching segments: Custom vs full contig
Actually represent the difference between a full contig that we loaded
and a subset of it that the user chose. This allows us to add the real
"last position" to the type.

Note: We don't use that yet -- this is just the refactor.

For convenience, this also adds a couple Deref impls and uses them to
make the calling code cleaner.