Report Genotype likelihood and confidence in Phred
Also fixes a mixup in `variant_calling.rs` where we gave the
phred-scaled probability not the chance of error.

This changes how these are printed: Previously rastair wrote the precise
float, now it only writes the rounded integer representation. This means
there is a slight error added when converting.

Fixes #20