bed writer: Fix double-flush bug
This adds a comment on closing BED files and a test to ensure they are
valid according to `tabix`.

Previously, when closing a BED file, we called `flush` on the writer.
Then, if it is a bgzip writer, we called `closed` on it to maybe get an
index file.

There seems to be a. bug it the bgzip crate: When we flush and then
close, the last block is written twice. This shows up when calling tabix
on it, since the block is not just out of order but also cut off at the
beginning.