bed: existing CpG > de-novo flag
In the test data, you can find a nice edge case in
`bacteriophage_lambda_CpG:42237`: The second C in a CCGG has a G alt,
forming a de-novo CpG candidate in the first two bases next to the ref
CpG in base 2 and 3.

In the BED output, we added a column that indicates de-novo CpGs, but in
cases like this it should not say this is new but rather call it REF.

There is still a hidden bug here: The first C is part of that de-novo
candidate, but we're not calling at all.