metrics2: fix methylation based on GT
Don't be clever. That's one of the rules I try to follow. If I don't,
stuff like this happens. We tried to call methylation but it depended
on the genotype being called before. But since we don't update the
metrics struct until all extended fields are calculated, it was never
read. Great job, Pascal.